Add additional terms
As well as routine additions and corrections these commit adds support for for multiple codes (e.g. different codes for the same term for male/female), and CodingFlags. It adds many metastatic codes using the extension code suggested by WHO, but we know that this will need to be replaced because its understanding of metastatic differs to that of certain (mostly English-speaking) countries. It also reflects a change underway to make 'carcinoma' an 'other specified' rather than 'unspecified' condition (i.e. a Y rather than a Z)